// xeloba.ge/src/Components/Interfaces/CraftsmanDataInterface.ts
export interface CraftsmanDataInterface{
    userID: string;
    setDate:string;
    name: string;
    lastname: string;
    personalID: string;
    email: string;
    phone_number: string;
    password: string;
    city: string;
    district: string;
    status: string;
    profession: string;
    experience: string;
    price: string;
    visitPrice: string;
    moreInfo: string;
    profileImgs: string;
    last_login: string;
}